# specs/001-simplify-rental-accessories/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 租赁管理系统 - 简化附件选择 API
  description: |
    租赁订单附件管理API规范,支持配套附件(手柄、镜头支架)和库存附件(手机支架、三脚架)的简化选择。
    
    **变更说明**:
    - 添加 `includes_handle` 和 `includes_lens_mount` 布尔字段
    - `accessories` 字段现在只包含库存附件(手机支架、三脚架)
    - 新增 `all_accessories` 字段返回完整附件信息(用于打印和甘特图)
  version: 2.0.0
  contact:
    name: 开发团队
servers:
  - url: http://localhost:5000/api
    description: 本地开发环境
  - url: https://api.inventory.example.com/api
    description: 生产环境
...
paths:
...
  /gantt/data:
    get:
      tags:
        - Gantt
      summary: 获取甘特图数据
      description: 获取指定时间范围内的租赁订单数据,用于甘特图渲染
      parameters:
        - name: start_date
          in: query
          required: true
          description: 查询开始日期(YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          description: 查询结束日期(YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: include_accessories
          in: query
          description: 是否包含附件信息
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GanttDataResponse'
...
components:
  schemas:
...
    FullAccessoryInfo:
      type: object
      properties:
        id:
          type: integer
          description: 附件设备ID(配套附件无此字段)
          example: 45
        name:
          type: string
          description: 附件名称
          example: "手柄"
        serial_number:
          type: string
          description: 附件序列号(配套附件无此字段)
          example: "P01-20240501"
        type:
          type: string
          enum: [phone_holder, tripod, handle, lens_mount]
          description: 附件类型
          example: "handle"
        is_bundled:
          type: boolean
          description: 是否为配套附件
          example: true
...
    GanttDataResponse:
      type: object
      properties:
        devices:
          type: array
          description: 设备列表(甘特图行)
          items:
            type: object
            properties:
              device_id:
                type: integer
              device_name:
                type: string
              device_model:
                type: string
        rentals:
          type: array
          description: 租赁记录(甘特图条)
          items:
            type: object
            properties:
              id:
                type: integer
              device_id:
                type: integer
              customer_name:
                type: string
              start_date:
                type: string
                format: date
              end_date:
                type: string
                format: date
              status:
                type: string
              accessories:
                type: array
                description: 附件信息(含配套附件和库存附件)
                items:
                  $ref: '#/components/schemas/FullAccessoryInfo'
...